requirements.txt
aiofiles==23.2.1
aiogram==3.15.0
aiohttp==3.9.5
aiosignal==1.3.1
amqp==5.3.0
annotated-types==0.7.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
billiard==4.2.1
cachetools==5.3.3
celery==5.4.0
certifi==2024.7.4
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.5
Deprecated==1.2.14
diff-match-patch==20241021
distro==1.9.0
dj-database-url==2.2.0
dj-email-url==1.0.6
Django==5.0.6
django-cache-url==3.4.5
django-celery-beat==2.7.0
django-ckeditor==6.7.1
django-import-export==4.2.1
django-jazzmin==3.0.0
django-js-asset==2.2.0
django-timezone-field==7.0
environs==11.0.0
ffmpeg==1.4
frozenlist==1.4.1
google-auth==2.31.0
google-auth-oauthlib==1.2.0
googletrans==4.0.0rc1
gspread==6.1.2
gTTS==2.5.4
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.12.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
kombu==5.4.2
magic-filter==1.0.12
marshmallow==3.21.3
multidict==6.0.5
netaddr==1.3.0
numpy==2.0.0
oauthlib==3.2.2
orjson==3.10.6
packaging==24.1
pandas==2.2.2
pillow==10.4.0
prompt_toolkit==3.0.48
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.7.4
pydantic_core==2.18.4
pydub==0.25.1
pytesseract==0.3.13
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
redis==5.2.0
requests==2.32.3
requests-oauthlib==2.0.0
rfc3986==1.5.0
rsa==4.9
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
SpeechRecognition==3.12.0
sqlparse==0.5.0
tablib==3.7.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
wrapt==1.16.0
yarl==1.9.4
yookassa==3.3.0
